+++ Extensions/IISWebAppDeploy/Tests/Tasks/SqlDacpacDeploy/SqlPackageOnTargetMachines.Tests.ps1
@@ -441,7 +441,7 @@ Describe "Tests for Get-SqlPackageCmdArgs functionality" {
 
     Context "When target method is server and with windows authentication" {
 
-        $cmdArgs = Get-SqlPackageCmdArgs -dacpacFile "sample.dacpac" -targetMethod "server" -serverName "localhost" -databaseName "SampleDB" -authscheme "Windows Authentication"
+        $cmdArgs = Get-SqlPackageCmdArgs -dacpacFile "sample.dacpac" -targetMethod "server" -serverName "localhost" -databaseName "SampleDB" -authscheme "windowsAuthentication"
 
         It "Should contain targetmethod as server and no sqluser argument should be present" {
             ($cmdArgs.Contains('/SourceFile:"sample.dacpac" /Action:Publish /TargetServerName:"localhost" /TargetDatabaseName:"SampleDB"')) | Should Be $true
@@ -451,7 +451,7 @@ Describe "Tests for Get-SqlPackageCmdArgs functionality" {
 
     Context "When target method is server and with mixed mode authentication" {
 
-        $cmdArgs = Get-SqlPackageCmdArgs -dacpacFile "sample.dacpac" -targetMethod "server" -serverName "localhost" -databaseName "SampleDB" -authscheme "SQL Server Authentication"-sqlUsername "dummyuser" -sqlPassword "dummypassword"
+        $cmdArgs = Get-SqlPackageCmdArgs -dacpacFile "sample.dacpac" -targetMethod "server" -serverName "localhost" -databaseName "SampleDB" -authscheme "sqlServerAuthentication"-sqlUsername "dummyuser" -sqlPassword "dummypassword"
 
         It "Should contain targetmethod as server and sqluser argument should be present" {
             ($cmdArgs.Contains('/SourceFile:"sample.dacpac" /Action:Publish /TargetServerName:"localhost" /TargetDatabaseName:"SampleDB"')) | Should Be $true